yolo/requirements.txt
requests == 2.28.2
numpy==1.24.3
opencv_python==4.8.0.76
Pillow==10.1.0
torch==2.0.0a0+gitc263bd4
MarkupSafe==2.1.3
torchvision==0.15.1a0
pylibtiff==0.6.1
pytesseract==0.3.10
ninja==1.11.1
scikit_image==0.22.0
scipy == 1.11.4
easyocr==1.7.1
ultralytics==8.0.216
GitPython==3.1.40
